refactor: organise file_upload JS logic into folders
