+++ nodes/web/helpers/file_upload/main.js
@@ -1,0 +1,3 @@
+import { addUploadWidget, NODES } from "./utils.js";
+
+export { addUploadWidget, NODES };
